+++ HomeworkAssignment.DTOs/RequestDTOs/AssignmentRelated/RequestPartialAssignmentDto.cs
@@ -10,9 +10,9 @@ public class RequestPartialAssignmentDto
     public string? RepositoryUrl { get; set; }
     public DateTime? Deadline { get; set; } 
 
-    public int? MaxScore { get; set; } 
-    public int? MaxAttemptsAmount { get; set; } 
-    public int? Position { get; set; }
+    public ushort? MaxScore { get; set; } 
+    public ushort? MaxAttemptsAmount { get; set; } 
+    public ushort? Position { get; set; }
     
     public bool? IsPublished { get; set; } 
     
@@ -22,12 +22,12 @@ public class RequestPartialAssignmentDto
     public bool? AttemptTestsSectionEnable { get; set; }
     public bool? AttemptQualitySectionEnable { get; set; }
 
-    public int? AttemptCompilationMaxScore { get; set; }
-    public int? AttemptCompilationMinScore { get; set; }
+    public ushort? AttemptCompilationMaxScore { get; set; }
+    public ushort? AttemptCompilationMinScore { get; set; }
 
-    public int? AttemptTestsMaxScore { get; set; } 
-    public int? AttemptTestsMinScore { get; set; }
+    public ushort? AttemptTestsMaxScore { get; set; } 
+    public ushort? AttemptTestsMinScore { get; set; }
 
-    public int? AttemptQualityMaxScore { get; set; } 
-    public int? AttemptQualityMinScore { get; set; } 
+    public ushort? AttemptQualityMaxScore { get; set; } 
+    public ushort? AttemptQualityMinScore { get; set; } 
 }
