+++ HomeworkAssignment.DTOs/RequestDTOs/AssignmentRelated/RequestReorderAssignmentDto.cs
@@ -3,5 +3,5 @@
 public class RequestReorderAssignmentDto
 {
     public required Guid Id { get; set; }
-    public int Position { get; set; }
+    public ushort Position { get; set; }
 }
