+++ HomeworkAssignment.DTOs/RequestDTOs/AttemptRelated/RequestPartialAttemptDto.cs
@@ -5,6 +5,6 @@ public class RequestPartialAttemptDto
     public Guid UserId { get; set; }
     public Guid? AssignmentId { get; set; }
     
-    public int? Position { get; set; }
+    public ushort? Position { get; set; }
     public string? BranchName { get; set; }
 }
