+++ HomeworkAssignment.DTOs/RequestDTOs/ChapterRelated/RequestReorderChapterDto.cs
@@ -3,5 +3,5 @@
 public class RequestReorderChapterDto
 {
     public required Guid Id { get; set; }
-    public int Position { get; set; }
+    public ushort Position { get; set; }
 }
